refactor(fetch): separate files to enhance code readability
